requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
textstat==0.7.3
nltk==3.8.1
python-dateutil==2.8.2
lxml==4.9.3
urllib3==2.0.4
